Moved k-RR mechanism to staircase module, because k-RR is a special case of staircase mechanisms.
