+++ diffprivlib/mechanisms/__init__.py
@@ -28,8 +28,7 @@ from diffprivlib.mechanisms.gaussian import Gaussian, GaussianAnalytic, Gaussian
 from diffprivlib.mechanisms.geometric import Geometric, GeometricFolded, GeometricTruncated
 from diffprivlib.mechanisms.laplace import Laplace, LaplaceBoundedDomain, LaplaceBoundedNoise, LaplaceFolded,\
     LaplaceTruncated
-from diffprivlib.mechanisms.staircase import Staircase
+from diffprivlib.mechanisms.staircase import Staircase, KRR
 from diffprivlib.mechanisms.uniform import Uniform
 from diffprivlib.mechanisms.vector import Vector
 from diffprivlib.mechanisms.wishart import Wishart
-from diffprivlib.mechanisms.krr import KRR
