+++ diffprivlib/mechanisms/staircase.py
@@ -19,9 +19,11 @@
 The staircase mechanism in differential privacy.
 """
 from numbers import Real
+from math import e
 
 import numpy as np
 
+from diffprivlib.mechanisms.base import DPMechanism
 from diffprivlib.mechanisms.laplace import Laplace
 from diffprivlib.utils import copy_docstring
 
@@ -99,3 +101,80 @@ class Staircase(Laplace):
         return value + sign * ((1 - binary_rv) * ((geometric_rv + self.gamma * unif_rv) * self.sensitivity) +
                                binary_rv * ((geometric_rv + self.gamma + (1 - self.gamma) * unif_rv) *
                                             self.sensitivity))
+
+class KRR(DPMechanism):
+    r"""
+    The k-RR mechanism - also known as flat mechanism or randomized response
+    mechanism - is a staircase mechanism with Y = X. It can be viewd as a
+    generalization of the randomised response proposed by [Warner, 1965].
+  
+    The k-RR mechahism returns the original value of an attribute domain with
+    probability :math:`\frac{e^\varepsilon}{|X|-1+e^\varepsilon}` and returns
+    any other value different from the original with probability
+    :math:`\frac{1}{|X|-1+e^\varepsilon}`, where :math:`|X|` is the domain
+    size. P. Kairouz, S. Oh, and P. Viswanath has showed that the mechanism is
+    optimal in the low privacy regime for a large class of information
+    theoretic utility functions.
+
+    Parameters
+    ----------
+    epsilon : float
+        Privacy parameter :math:`\epsilon` for the mechanism. Must be in [0, ∞).
+    
+    domain_values : list
+        Domain of values of the mechanism.
+    
+    """
+
+    def __init__(self, *, epsilon, domain_values):
+        super().__init__(epsilon=epsilon, delta=0.0)
+        self._domain_values, self._data_size = self._check_domain(domain_values)
+
+    def _check_domain(self, domain_values):
+        if not isinstance(domain_values, list):
+            raise TypeError("Domain must be a list")
+        
+        if len(domain_values) <= 1:
+            raise ValueError("The domain must have at least 2 elements")
+
+        return domain_values, len(domain_values)
+
+    @copy_docstring(DPMechanism.bias)
+    def bias(self, value):
+        raise NotImplementedError
+
+    @copy_docstring(DPMechanism.variance)
+    def variance(self, value):
+        raise NotImplementedError
+
+    def randomise(self, value):
+        """Randomise `value` with the mechanism.
+
+        Parameters
+        ----------
+        value : int or other
+            The value to be randomised.
+
+        Returns
+        -------
+        int or other
+            The randomised value, same type as `value`.
+
+        """
+
+        self._check_all(value)
+        unif_prob = self._rng.random()
+
+        if (e**self.epsilon == float("inf") or 
+            unif_prob <= e**self.epsilon / 
+            (self._data_size - 1 + e**self.epsilon)):
+            return value
+        
+        """Select a random index from domain that is different from the index
+        of the value that is being randomised
+        """
+        idx = self._rng.randrange(self._data_size)
+        while self._domain_values[idx] == value:
+            idx = self._rng.randrange(self._data_size)
+
+        return self._domain_values[idx]
